Separates tableau from its state
